Move BlockValidator to separate file
This does not make any changes to the BlockValidator class
